Per Issue #1: - Added support for python 3 (tested with 3.6.1) - Broke dynamic database functionality out of the model and into it's own class.   The model is a method to store a config for a DynamicDatabase instance but you   no longer need a model instance to get a dynamic database instance.   You can instantiate it yourself with a name and database config dictionary,   identical to what you set in the settings.py file.
