+++ dynamic_databases/__init__.py
@@ -1,4 +1,4 @@
 
-__version__ = '0.1.6'
+__version__ = '0.1.7'
 
 default_app_config = 'dynamic_databases.apps.DynamicDatabasesConfig'
